refactor: components -> feature
