+++ src/app/layout.tsx
@@ -3,8 +3,8 @@ import { notoSansJP } from "@/lib/metadata";
 import { ThemeProvider } from "next-themes";
 import Footer from "./footer";
 import "./globals.css";
-import { ReferralCodeHandlerWrapper } from "@/components/ReferralCodeHandlerWrapper";
 import { Toaster } from "@/components/ui/sonner";
+import { ReferralCodeHandlerWrapper } from "@/features/referral/components/referral-code-handler-wrapper";
 import { generateRootMetadata } from "@/lib/metadata";
 import Script from "next/script";
 import { Suspense } from "react";
