+++ src/features/mission-detail/components/mission-form-wrapper.tsx
@@ -2,7 +2,6 @@
 
 import { SubmitButton } from "@/components/submit-button";
 import { Button } from "@/components/ui/button";
-import { XpProgressToastContent } from "@/components/xp-progress-toast-content";
 import { achieveMissionAction } from "@/features/mission-detail/actions/actions";
 import { MainLinkButton } from "@/features/mission-detail/components/main-link-button";
 import { MissionCompleteDialog } from "@/features/mission-detail/components/mission-complete-dialog";
@@ -10,6 +9,7 @@ import { ArtifactForm } from "@/features/missions/components/artifact-form";
 import QuizComponent from "@/features/missions/components/quiz-component";
 import { useMissionSubmission } from "@/features/missions/hooks/use-mission-submission";
 import { useQuizMission } from "@/features/missions/hooks/use-quiz-mission";
+import { XpProgressToastContent } from "@/features/user-level/components/xp-progress-toast-content";
 import { ARTIFACT_TYPES } from "@/lib/types/artifact-types";
 import type { Tables } from "@/lib/types/supabase";
 import type { User } from "@supabase/supabase-js";
